MAINT: Move fixture from factories.py to fixtures.py
